Android Release v0.10.6

+++ wallet/src/de/schildbach/wallet/ui/SendingAddressesFragment.java
@@ -196,15 +196,13 @@ public final class SendingAddressesFragment extends SherlockListFragment impleme
 	@Override
 	public boolean onOptionsItemSelected(final MenuItem item)
 	{
-		switch (item.getItemId())
-		{
-			case R.id.sending_addresses_options_paste:
-				handlePasteClipboard();
-				return true;
-
-			case R.id.sending_addresses_options_scan:
-				handleScan();
-				return true;
+		int itemid = item.getItemId();
+		if( itemid == R.id.sending_addresses_options_paste){
+			handlePasteClipboard();
+			return true;
+		}else if( itemid == R.id.sending_addresses_options_scan){
+			handleScan();
+			return true;
 		}
 
 		return super.onOptionsItemSelected(item);
@@ -286,39 +284,34 @@ public final class SendingAddressesFragment extends SherlockListFragment impleme
 			@Override
 			public boolean onActionItemClicked(final ActionMode mode, final MenuItem item)
 			{
-				switch (item.getItemId())
-				{
-					case R.id.sending_addresses_context_send:
-						handleSend(getAddress(position));
+				
+				int itemid = item.getItemId();
+				if(itemid == R.id.sending_addresses_context_send){
+					handleSend(getAddress(position));
 
-						mode.finish();
-						return true;
+					mode.finish();
+					return true;
+				}else if(itemid == R.id.sending_addresses_context_edit ){
+					EditAddressBookEntryFragment.edit(getFragmentManager(), getAddress(position));
 
-					case R.id.sending_addresses_context_edit:
-						EditAddressBookEntryFragment.edit(getFragmentManager(), getAddress(position));
+					mode.finish();
+					return true;	
+				}else if(itemid == R.id.sending_addresses_context_remove){
+					handleRemove(getAddress(position));
 
-						mode.finish();
-						return true;
+					mode.finish();
+					return true;	
+				}else if(itemid == R.id.sending_addresses_context_show_qr){
+					handleShowQr(getAddress(position));
 
-					case R.id.sending_addresses_context_remove:
-						handleRemove(getAddress(position));
+					mode.finish();
+					return true;	
+				}else if(itemid == R.id.sending_addresses_context_copy_to_clipboard){
+					handleCopyToClipboard(getAddress(position));
 
-						mode.finish();
-						return true;
-
-					case R.id.sending_addresses_context_show_qr:
-						handleShowQr(getAddress(position));
-
-						mode.finish();
-						return true;
-
-					case R.id.sending_addresses_context_copy_to_clipboard:
-						handleCopyToClipboard(getAddress(position));
-
-						mode.finish();
-						return true;
+					mode.finish();
+					return true;	
 				}
-
 				return false;
 			}
